Ensure the implementation of Eigen spatial convolutions in TF and TFLite in sync.
PiperOrigin-RevId: 237958691
